+++ js/util.js
@@ -1,0 +1,32 @@
+export const getRandomNumber = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+export const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export function createRandomIdFromRangeGenerator(min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomNumber(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomNumber(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export function generateText(sentence) {
+  return function () {
+    const getSentenceId = createRandomIdFromRangeGenerator(0, sentence.length - 1);
+    return sentence[getSentenceId()];
+  };
+}
